ucode: remove redundant ->priv null pointer checks
The ucv_get() and ucv_put() functions are NULL safe, so drop various
redundant checks guarding calls to them.

Also always push the ->priv value onto the stack and invoke the user
callbacks with the appropriate amounts of arguments. Attempts to
read unset priv data in managed code will yield `null` is either case,
regardless of whether NULL is pushed onto the stack or whether the
callback function is invoked with one less argument.

This slightly reduces the complexity of the invocation code.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>